Changing check to avoid duplicate points
I noticed that having duplicate points in the dataset caused those points to be counted multiple times.  The solution is to remove the duplicate points.  To check if a point has been added, I was checking to see if the (x, 'p') tuple was in the list of all points.  However, that is VERY slow.  It was much faster to just check if the point value, x, had been added to the points lookup dictionary yet.  This decreased run time from ~40 seconds for a large dataset to instant.